FN-11336 Plugin refactoring. Make main module startup script compatible with PHP 5.6, added PHP version checking at startup and added informational messages if environment does not meet minimal requirements to make module more stable, reliable and user-friendly.

+++ upgrade/init-3.0.0.php
@@ -27,8 +27,15 @@ if (!defined('_PS_VERSION_')) {
     exit;
 }
 
-function upgrade_module_3_0_0(Comfino $module): bool
+/**
+ * @return bool
+ */
+function upgrade_module_3_0_0(Comfino $module)
 {
+    if (!$module->checkEnvironment()) {
+        return false;
+    }
+
     if (file_exists(_PS_MODULE_DIR_ . 'comfino/override')) {
         // Remove admin controller override from older versions.
         unlink(_PS_MODULE_DIR_ . 'comfino/override/index.php');
